Initial action manager object
Add an action manager object that ensures that actions run, honoring
prerequsite order, grouping and run_result.

The bulk of the work is in the "sorting" of actions to preserve
prerequsite order. The comparison function checks to see if one is a
prerequsite of another, either directly or recursively.

The actions are run via a run() method, which iterates over groups,
finds the actions for each group, sorts the actions by prerequsites,
and calls each action's run() method.

Based on the run_result, the manager will:
- proceed with no errors
- stop with no errors
- print any returned error string.

The run result from the last action is returned from run().

After all groups have run, all groups are deinit()ed using each
action's deinit() method.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>