new(engine): add classes to resolve list and macro references in rule filter
This is a first step towards porting the rule filter building logic that is currently implemented in Lua.
filter_list_resolver and filter_macro_resolver use the newly introduced AST constructs from libsinsp, and
allow manipulating filter ASTs to resolve/replace list and macro references. This is meant to be used
at boot time by the rule loader (which we'll still maintain implemented in Lua for now).

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>